Proposed fix for issue #2
After the user reported that the last fix still led to a crash, I worked to include a check to see if the list was empty, and if so, just return, otherwise, work to extract the string. This is more than likely some other version of SNMP or something else entirely that does not use a single ascii printable character within the data packet.