MAC address OUI VM detection and UDP broadcast service detection.
MAC address OUI VM based detection attempts to use the OUI to determine if the target host is a physical or virtual machine, and if a VM what kind of VM it might consist of. The second update has to do with UDP broadcast traffic. This really impacts the way Prebellico reports broadcast traffic. If it finds a new host through broadcast, the message is slightly different. If it notices a host continually broadcasting traffic from a certain port, it starts to track it and alerts the user. If this traffic continues, Prebellico considers it an open UDP port, records it in the database and no longer reports the traffic. This last update ensures that something that if something is continually broadcasting UDP traffic is not filling the Prebellico log with the same message.