New basic forms of authenitcation detection and processing.
Added support for clear-text protocols including things like FTP/Telnet/Basic Authentication/etc. Also added support for NTLM authentication. Evidence shows this is also leaked. Just need to find a way to extract hashes out the packets. Turns out this is not as easy as I had hoped. Note that none of these new methods of detection are committed to the database.